requirements.txt
PyQt6>=6.6.0
dotenv>=0.9.9
mysql-connector-python==9.0.0